Added second interrupt handler.  The original handler is for rising signal interrupts which occur when a button is pressed. The new handler is for falling signal interrupts which occur when a button is released.
The new behavior is that once a pressed button interrupt has been handled, the code will ignore all other pressed interrupts
until a released button interrupt has occurred.  There is bounce occurring when the button is pressed and released.  This
prevents secondary pressed button interrupts that occur after the first one is detected and before the first one is processed
from triggering false detection.  Again this is due to bounce and the fact that the Pico does not process an interrupt
immediately and schedules the handler which allows multiple interrupts to trigger handling for the same button press.